requirements.txt
# git+ssh://git@github.com/dktunited/discoveryit-airflow-discovery-operators.git
-e dps-airflow-operators/
apache-airflow[ssh]==2.2.2
apache-airflow-providers-sftp==2.2.0
hvac==0.11.2
PyYAML
anyconfig==0.12.0
